requirements.txt
numpy
tqdm
pytorch
torchvision
pytorch3d
matplotlib
tensorboard
gym
dm_control
ruamel.yaml